Add #full_attachment_path and make sure it doesn't get deleted even when empty.
This is because on production apps, it's very common for the attachment_path to be a symlink to a shared dir (like public/files => shared/public/files), but attachment_fu was removing the link whenever it was deleting the last attachment file, or even updating it. So the symlink was being deleted and new attachments where being saved directly in a newly created (by attachment_fu) public/files directory in the release path instead of the shared path.